Support for the --all flag
Allow setting the --all/-a flag, which means that dot-files are
not filtered in the directory walking. This additionally leads
us to be forced to deal with the ".git/" directory, one which
git by contract refused to parse. (Try `git add .git` to see.)
As such, when the all flag is set, we instantiate an override
for our walker.

Additionally, this changeset performs a little bit of tidying
for returned results, following the general structure of
`greprs` from the Rust book, which presumably reflects best
practices.